+++ adx.js
@@ -24,7 +24,7 @@ JSE.jseTestNet = commandLine.testnet;
 
 if (JSE.jseTestNet !== false) console.log('WARNING: RUNNING IN TESTNET MODE - '+JSE.jseTestNet); // idiot check
 
-JSE.jseVersion = 'JSEcoin Ad Exchange v1.9.01';
+JSE.jseVersion = 'JSEcoin Ad Exchange v1.9.1';
 
 const fs = require('fs');
