+++ datastore.js
@@ -4,7 +4,7 @@
  * @example forever start -c "node --max-old-space-size=11500" datastore.js &
  * @example forever start -c "node --max-old-space-size=10000" datastore.js -f blockChain &
  * @example node --max-old-space-size=3000 datastore.js -p 82 -t local
- * @version 1.8.2
+ * @version 1.9.1
  * @description The JSE datastore is a custom key value storage system design to efficiently handle Javascript objects via socket.io
  */
 
@@ -31,7 +31,7 @@ const negFilter = commandLine.negfilter;
 
 if (JSE.jseTestNet !== false) console.log('WARNING: RUNNING IN TESTNET MODE - '+JSE.jseTestNet); // idiot check
 
-JSE.jseVersion = 'JSEcoin Datastore Server v1.8.2';
+JSE.jseVersion = 'JSEcoin Datastore Server v1.9.1';
 
 const dataDir = './data/'; // start with ./ relative to this files location ane end in slash i.e. ./data/
 const bkupDir = './logs/'; // as above
